third_party/WebKit/LayoutTests/fast/js/Promise-init.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

description('Test Promise.');

var thisInInit;
var resolver;
var promise = new Promise(function(r) {
  thisInInit = this;
  resolver = r;
});

shouldBeTrue('promise instanceof Promise');
shouldBe('promise.constructor', 'Promise');
shouldBe('thisInInit', 'promise');
shouldBeTrue('resolver instanceof PromiseResolver');
shouldBe('resolver.constructor', 'PromiseResolver');

shouldThrow('new Promise()', '"TypeError: Promise constructor takes a function argument"');
shouldThrow('new Promise(37)', '"TypeError: Promise constructor takes a function argument"');

shouldNotThrow('promise = new Promise(function() { throw Error("foo"); })');
// FIXME: We can't test if the promise is rejected now, but should test it in the future.

</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
